+++ lib/axn/webhooks.rb
@@ -5,6 +5,7 @@ require "active_support/deprecation"
 
 require_relative "webhooks/version"
 require_relative "webhooks/request"
+require_relative "webhooks/signature"
 
 module Axn
   module Webhooks